Squashed commit of the following:
commit 6289e7f8de432ebdcf8c9027a4ebc09997e24a57
Author: Zikani Nyirenda Mwase <zikaninyirenda@cgatechnologies.org.uk>
Date:   Tue Mar 1 11:53:05 2022 +0200

    fix: Update the query the for fetching households

commit 6eadcaf70477bc007455547b4db2f1e25d4bd4e3
Author: Zikani Nyirenda Mwase <zikaninyirenda@cgatechnologies.org.uk>
Date:   Tue Mar 1 11:52:27 2022 +0200

    fix: Transfer agency assignment form

commit 635e4ab1982118354efde8f8783a498c4041db11
Author: Zikani Nyirenda Mwase <zikaninyirenda@cgatechnologies.org.uk>
Date:   Tue Mar 1 11:51:56 2022 +0200

    fix: Add axios library for Ajax requests

commit ee1172c6b898d7bdc3c9fdb584924600d93bed61
Author: Zikani Nyirenda Mwase <zikaninyirenda@cgatechnologies.org.uk>
Date:   Tue Mar 1 11:49:46 2022 +0200

    fix: Fix data types and transfer agency assignment

commit a53ae87470e154b1ca977eae8240db0f179a8d97
Author: Zikani Nyirenda Mwase <zikaninyirenda@cgatechnologies.org.uk>
Date:   Tue Mar 1 11:48:29 2022 +0200

    fix: Fix edit pages for parameters controllers

+++ sctp-core/src/main/java/org/cga/sctp/transfers/TransferEvent.java
@@ -113,7 +113,7 @@ public class TransferEvent {
     private Long totalTransferAmount;// valuetotaltransfer;
 
     @Column
-    private Long totalMembers;
+    private Integer totalMembers;
 
     @Column
     private Long totalMembersPrimary;
@@ -342,11 +342,11 @@ public class TransferEvent {
         this.totalTransferAmount = totalTransferAmount;
     }
 
-    public Long getTotalMembers() {
+    public Integer getTotalMembers() {
         return totalMembers;
     }
 
-    public void setTotalMembers(Long totalMembers) {
+    public void setTotalMembers(Integer totalMembers) {
         this.totalMembers = totalMembers;
     }
